Move batch controller into a separate action
