+++ server/actions/index.js
@@ -1,1 +1,2 @@
 export getContactProperties from "./get-contact-properties";
+export handleBatch from "./handle-batch";
