+++ server/worker.js
@@ -3,7 +3,6 @@ import { Connector } from "hull";
 
 import AppMiddleware from "./lib/middleware/app";
 
-import BatchController from "./controller/batch";
 import MonitorController from "./controller/monitor";
 import UsersController from "./controller/users";
 import FetchAllController from "./controller/fetch-all";
@@ -12,7 +11,6 @@ import * as newJobs from "./jobs";
 
 export default function workerJobs(connector: Connector) {
   const jobs = {
-    handleBatchExtractJob: BatchController.handleBatchExtractJob,
     fetchAllJob: FetchAllController.fetchAllJob,
     saveContactsJob: UsersController.saveContactsJob,
     sendUsersJob: UsersController.sendUsersJob,
